requirements.txt
# FastAPI and server (updated for CS platform - v2.1)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.0

# Validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client and forms
httpx>=0.26.0
aiohttp>=3.9.0
python-multipart>=0.0.6

# Twilio
twilio>=8.10.0

# Email (SendGrid)
sendgrid>=6.10.0

# Background Job Scheduler
apscheduler>=3.10.0

# Auth - pinned to avoid CVE-2024-23342 in ecdsa
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
# Pin cryptography to get ecdsa fix
cryptography>=42.0.0

# Validation
email-validator>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=22.0.0

# OpenTelemetry (optional - for distributed tracing)
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-instrumentation-redis>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0

# Redis (optional - for distributed caching)
redis>=5.0.0

# Development
python-dotenv>=1.0.0

# PDF Generation for Inspection Letters
weasyprint>=60.0
Jinja2>=3.1.0

# Web scraping for TDEC permit lookup
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Force rebuild Wed, Jan  7, 2026  9:32:51 AM
# Build trigger: 1767807723
